[giac] Updating newest files
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@24698 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/giac/moyal.h
@@ -24,7 +24,7 @@
 namespace giac {
 #endif // ndef NO_NAMESPACE_GIAC
   class gen;
-  gen incomplete_beta(int a,int b,double p,bool regularize=true);
+  gen incomplete_beta(double a,double b,double p,bool regularize=true);
   gen lower_incomplete_gamma(double s,double z,bool regularize=true); // lower incomplete
